[Refactoring] Move MoveMembersToExtension to a separate file
